%p printed half of every pointer, and the log was the last place for that
    before:  MEM ALLOC 18(96) - 0x86aaefa0
    after:   MEM ALLOC 18(96) - 0xfffffabb86aaefa0

common/snprintf.c fetched %p as (unsigned long), which is four bytes under
LLP64. So every pointer this driver has ever logged arrived with its top half
missing -- kernel addresses on this host are 0xfffffabb........ and
0xfffffadf........, and what came out was the low word alone. Three call
sites, one per prefix state; the accumulator was already unsigned long long,
so only the casts were wrong.

This is the same truncation class the port has been finding since the ELF
loader (vm_ptr_t, the interrupt.c ISR address, CO_ARCH_PAGE_MASK), and this
instance was the expensive one, because it was in the instrument. A pool
address printed as 0x86aaefa0 cannot be compared against a bugcheck
parameter of 0xfffffadfbe266892, cannot be matched to an allocation, and
cannot be recognised as belonging to one region rather than another -- and it
does not look broken, it looks like a small address. Tonight's investigation
spent hours hand-decoding minidumps while the driver's own log, had anyone
been able to read it, was printing the answer with the identifying half
removed.

There is a commit in this history about exactly this, from the ud2 stepping
work: an instrument that turns a specific answer into a vague one is worse
than no instrument, because the vague answer gets believed. It applied to a
fault report then and to every pointer in the log now.

In fairness to 2003: unsigned long held a pointer on every machine that code
ran on. LLP64 made it a lie, which makes it this port's inheritance rather
than anyone's carelessness.

+++ src/colinux/common/snprintf.c
@@ -514,7 +514,7 @@ static int pvsnfmt_int(char **pinsertion, long *nmax, char fmt, int flags,
                 numbersigned = 0;
                 break;
              case 'p':
-                unumber = (unsigned long) va_arg(*ap, void *);
+                unumber = (unsigned long long)(size_t) va_arg(*ap, void *);
                 numbersigned = 0;
         }
         break;
@@ -552,7 +552,7 @@ static int pvsnfmt_int(char **pinsertion, long *nmax, char fmt, int flags,
                 numbersigned = 0;
                 break;
              case 'p':
-                unumber = (unsigned long) va_arg(*ap, void *);
+                unumber = (unsigned long long)(size_t) va_arg(*ap, void *);
                 numbersigned = numbersigned;
         }
         break;
@@ -571,7 +571,7 @@ static int pvsnfmt_int(char **pinsertion, long *nmax, char fmt, int flags,
                 numbersigned = 0;
                 break;
              case 'p':
-                unumber = (unsigned long) va_arg(*ap, void *);
+                unumber = (unsigned long long)(size_t) va_arg(*ap, void *);
                 numbersigned = 0;
          }
     } /* switch fmt to retrieve number */
